# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.00"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthResponse"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthResponse"
                    },
                    "422": {
                        "description": "User with same email already exists"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BaseResponse": {
                "description": "base response",
                "properties": {
                    "result": {
                        "title": "Result object",
                        "type": "object"
                    },
                    "message": {
                        "title": "Error message",
                        "type": "string"
                    },
                    "errors": {
                        "title": "Errors array",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User resource",
                "properties": {
                    "uuid": {
                        "title": "UUID",
                        "description": "User UUID",
                        "type": "string",
                        "example": "d1b0c6be-e64d-4baf-acec-17e7c6162216"
                    },
                    "name": {
                        "title": "Name",
                        "description": "User name",
                        "type": "string",
                        "example": "Ilya"
                    },
                    "email": {
                        "title": "Email",
                        "description": "User email",
                        "type": "string",
                        "example": "test@example.com"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "AuthResponse": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/components/schemas/BaseResponse"
                                },
                                {
                                    "properties": {
                                        "result": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "1|xw52hNtrkR1FGM3PcAPMmf1nhY1esAkJYP56ZIuV"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            ]
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "access_token for access to /user and /cars",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}